lsp: Remove filepath provider, move code into util module
Rather than have a provider dedicated to all things filepaths this
commit removes the existing filepath provider and creates a utility
module that exposes the following functions

- `complete_filepaths`: For providing completion suggestions for files
  relative to the current document.
- `complete_sphinx_filepaths`: For providing completion suggestions that
  match Sphinx's behaviour i.e. paths that begin with a `/` are relative
  to the project's `srcdir`
- `path_to_completion_item`: For taking a `pathlib.Path` and creating a
  `CompletionItem` from it. This function also contains the logic for
  constructing a `TextEdit` that correctly takes into account existing
  path segments.

This should then let us transition to creating providers decdicated to
particular roles/directives, opening up the possibility of them further
specialising later on.
