+++ lib/esbonio/esbonio/lsp/sphinx/__init__.py
@@ -51,7 +51,6 @@ DEFAULT_MODULES = [
     "esbonio.lsp.roles",
     "esbonio.lsp.sphinx.codeblocks",
     "esbonio.lsp.sphinx.domains",
-    "esbonio.lsp.sphinx.filepaths",
 ]
 """The modules to load in the default configuration of the server."""
